Clone ptero from within VM
Because we relied on ptero being cloned outside of the VM, the host OS
was able to do nasty things like insert CRLF's at the end of every line
(on windows for example).  This should make the VM more stand-alone.  If
you want to avoid the extra 'git clone' then supply PTERO_CHECKOUT as a
relative or absolute path to the ptero checkout you want to use.